[#2224] Finish refactoring workshops.util into other files
