+++ amy/extrequests/views.py
@@ -64,8 +64,8 @@ from workshops.models import (
     TrainingRequest,
     WorkshopRequest,
 )
-from workshops.util import merge_objects
 from workshops.utils.access import OnlyForAdminsMixin, admin_required
+from workshops.utils.merge import merge_objects
 from workshops.utils.metadata import fetch_workshop_metadata, parse_workshop_metadata
 from workshops.utils.trainingrequest_upload import (
     clean_upload_trainingrequest_manual_score,
