+++ amy/workshops/consts.py
@@ -1,1 +1,7 @@
 FEE_DETAILS_URL = "https://carpentries.org/workshops/#workshop-cost"
+
+STR_SHORT = 10  # length of short strings
+STR_MED = 40  # length of medium strings
+STR_LONG = 100  # length of long strings
+STR_LONGEST = 255  # length of the longest strings
+STR_REG_KEY = 20  # length of Eventbrite registration key
