+++ amy/workshops/models.py
@@ -34,8 +34,15 @@ from social_django.models import UserSocialAuth
 
 from autoemails.mixins import RQJobsMixin
 from workshops import github_auth
-from workshops.consts import FEE_DETAILS_URL
-from workshops.fields import NullableGithubUsernameField
+from workshops.consts import (
+    FEE_DETAILS_URL,
+    STR_LONG,
+    STR_LONGEST,
+    STR_MED,
+    STR_REG_KEY,
+    STR_SHORT,
+)
+from workshops.fields import NullableGithubUsernameField, choice_field_with_other
 from workshops.mixins import (
     ActiveMixin,
     AssignmentMixin,
@@ -53,12 +60,7 @@ from workshops.mixins import (
 )
 from workshops.signals import person_archived_signal
 from workshops.utils.dates import human_daterange
-
-STR_SHORT = 10  # length of short strings
-STR_MED = 40  # length of medium strings
-STR_LONG = 100  # length of long strings
-STR_LONGEST = 255  # length of the longest strings
-STR_REG_KEY = 20  # length of Eventbrite registration key
+from workshops.utils.emails import find_emails
 
 # ------------------------------------------------------------
 
@@ -1562,8 +1564,6 @@ class Event(AssignmentMixin, RQJobsMixin, models.Model):
     def mailto(self):
         """Return list of emails we can contact about workshop details, like
         attendance."""
-        from workshops.util import find_emails
-
         emails = find_emails(self.contact)
         return emails
 
@@ -1920,9 +1920,6 @@ class TrainingRequest(
     SecondaryEmailMixin,
     models.Model,
 ):
-
-    from workshops.util import choice_field_with_other
-
     MANUAL_SCORE_UPLOAD_FIELDS = (
         "request_id",
         "score_manual",
