+++ amy/workshops/utils/merge.py
@@ -1,84 +1,8 @@
-from collections import defaultdict
-import datetime
-import re
-
-from django.conf import settings
 from django.contrib.contenttypes.models import ContentType
-from django.contrib.sites.models import Site
-from django.db import IntegrityError, models, transaction
-from django.utils import timezone
+from django.db import IntegrityError, transaction
 from django_comments.models import Comment
 
-from consents.models import Consent
-from dashboard.models import Criterium
-from workshops.models import STR_LONG, STR_MED
-
-WORD_SPLIT = re.compile(r"""([\s<>"']+)""")
-SIMPLE_EMAIL = re.compile(r"^\S+@\S+\.\S+$")
-
-
-def find_emails(text):
-    """Find emails in the text.  This is based on Django's own
-    `django.utils.html.urlize`."""
-    # Split into tokens in case someone uses for example
-    # 'Name <name@gmail.com>' format.
-    emails = []
-
-    for word in WORD_SPLIT.split(text):
-        if SIMPLE_EMAIL.match(word):
-            local, domain = word.rsplit("@", 1)
-            try:
-                domain = domain.encode("idna").decode("ascii")
-            except UnicodeError:
-                continue
-            emails.append("{}@{}".format(local, domain))
-
-    return emails
-
-
-def archive_least_recent_active_consents(object_a, object_b, base_obj):
-    """
-    There is a unique database constraint on consents that only allows
-    (person, term) when archived_at is null.
-
-    This method archives one of the two active terms so
-    that the combine merge method will be successful.
-    """
-    consents = Consent.objects.filter(person__in=[object_a, object_b])
-    # Identify and group the active consents by term id
-    active_consents_by_term_id = defaultdict(list)
-    for consent in consents:
-        if consent.archived_at is None:
-            active_consents_by_term_id[consent.term_id].append(consent)
-
-    # archive least recent active consents
-    consents_to_archive = []
-    consents_to_recreate = []
-    for term_consents in active_consents_by_term_id.values():
-        if len(term_consents) < 2:
-            continue
-        consent_a, consent_b = term_consents[0], term_consents[1]
-        if consent_a.created_at < consent_b.created_at:
-            consents_to_archive.append(consent_a)
-        elif consent_b.created_at < consent_a.created_at:
-            consents_to_archive.append(consent_b)
-        else:
-            # If they were created at the same time rather than being
-            # nondeterministic archive both and when the user logs in again
-            # they can consent to the term once more.
-            consents_to_archive.append(consent_a)
-            consents_to_archive.append(consent_b)
-            consents_to_recreate.append(
-                Consent(
-                    person=base_obj,
-                    term=consent_a.term,
-                    term_option=None,
-                )
-            )
-    Consent.objects.filter(pk__in=[c.pk for c in consents_to_archive]).update(
-        archived_at=timezone.now()
-    )
-    Consent.objects.bulk_create(consents_to_recreate)
+from workshops.utils.consents import archive_least_recent_active_consents
 
 
 def merge_objects(
@@ -250,85 +174,3 @@ def merge_objects(
         merging_obj.delete()
 
         return base_obj.save(), integrity_errors
-
-
-def dict_without_Nones(**keys):
-    return {k: v for k, v in keys.items() if v is not None}
-
-
-def str2bool(v):
-    """Decoding string-encoded bool into Python objects.
-    This function returns one of the following:
-    * True if value decodes to true,
-    * False if value decodes to false,
-    * None otherwise.
-
-    Based on: https://stackoverflow.com/a/715468"""
-    value = str(v).lower()
-    if value in ("yes", "true", "t", "1"):
-        return True
-    elif value in ("no", "false", "f", "0"):
-        return False
-    else:
-        return None
-
-
-def choice_field_with_other(choices, default, verbose_name=None, help_text=None):
-    assert default in [c[0] for c in choices]
-    assert all(c[0] != "" for c in choices)
-
-    field = models.CharField(
-        max_length=STR_MED,
-        choices=choices,
-        verbose_name=verbose_name,
-        help_text=help_text,
-        null=False,
-        blank=False,
-        default=default,
-    )
-    other_field = models.CharField(
-        max_length=STR_LONG,
-        verbose_name=" ",
-        null=False,
-        blank=True,
-        default="",
-    )
-    return field, other_field
-
-
-def match_notification_email(obj):
-    """Try to match applied object to a set of criteria (defined in
-    `settings.py`)."""
-    results = []
-
-    # some objects may not have this attribute, in this case we should fall
-    # back to default criteria email
-    if hasattr(obj, "country") and obj.country:
-        results = Criterium.objects.filter(countries__contains=obj.country).values_list(
-            "email", flat=True
-        )
-    else:
-        # use general notification criteria when event has no country
-        results = [settings.ADMIN_NOTIFICATION_CRITERIA_DEFAULT]
-
-    # fallback to default address if nothing matches
-    return results or [settings.ADMIN_NOTIFICATION_CRITERIA_DEFAULT]
-
-
-def add_comment(content_object, comment, **kwargs):
-    """A simple way to create a comment for specific object."""
-    user = kwargs.get("user", None)
-    user_name = kwargs.get("user_name", "Automatic comment")
-    submit_date = kwargs.get(
-        "submit_date", datetime.datetime.now(datetime.timezone.utc)
-    )
-    site = kwargs.get("site", Site.objects.get_current())
-
-    return Comment.objects.create(
-        comment=comment,
-        content_object=content_object,
-        user=user,
-        user_name=user_name,
-        submit_date=submit_date,
-        site=site,
-    )
